fix error in PUT sfCAPE

+++ db/migrations/production/20190915221644_submissions_table_update.js
@@ -19,7 +19,11 @@ exports.down = function(knex) {
         return knex.schema.table("submissions", function(table) {
           table.dropColumn("submission_status");
           table.dropColumn("submission_errors");
-          table.dropColumn("card_brand");
+          knex.schema.hasColumn("card_brand").then(function(exists) {
+            if (exists) {
+              table.dropColumn("card_brand");
+            }
+          });
         });
       }
     })
